+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 - A bug when using an invalid nick in `/msg` command fixed (#111).
 - Bumped native-tls dependency -- fixes build for newer OpenSSLs (#114).
 - A bug when sending multi-line text (via C-x or pasting) fixed (#113).
+- Update to `C-x` (paste mode): empty lines are now sent as a space (" ").
+  Useful when e.g. sending long text with multiple paragraphs (#112).
 
 # 2018/09/01: 0.4.3